// frontend/src/api-client/models/TwitchClipsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reelshelf | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { TwitchClipSummary } from "./TwitchClipSummary";
import {
  TwitchClipSummaryFromJSON,
  TwitchClipSummaryFromJSONTyped,
  TwitchClipSummaryToJSON,
  TwitchClipSummaryToJSONTyped,
} from "./TwitchClipSummary";
import type { TwitchClipsState } from "./TwitchClipsState";
import {
  TwitchClipsStateFromJSON,
  TwitchClipsStateFromJSONTyped,
  TwitchClipsStateToJSON,
  TwitchClipsStateToJSONTyped,
} from "./TwitchClipsState";

/**
 *
 * @export
 * @interface TwitchClipsResponse
 */
export interface TwitchClipsResponse {
  /**
   *
   * @type {TwitchClipsState}
   * @memberof TwitchClipsResponse
   */
  state: TwitchClipsState;
  /**
   *
   * @type {string}
   * @memberof TwitchClipsResponse
   */
  twitchLogin: string | null;
  /**
   *
   * @type {Array<TwitchClipSummary>}
   * @memberof TwitchClipsResponse
   */
  clips: Array<TwitchClipSummary>;
  /**
   *
   * @type {string}
   * @memberof TwitchClipsResponse
   */
  cursor: string | null;
}

/**
 * Check if a given object implements the TwitchClipsResponse interface.
 */
export function instanceOfTwitchClipsResponse(
  value: object,
): value is TwitchClipsResponse {
  if (!("state" in value) || value["state"] === undefined) return false;
  if (!("twitchLogin" in value) || value["twitchLogin"] === undefined)
    return false;
  if (!("clips" in value) || value["clips"] === undefined) return false;
  if (!("cursor" in value) || value["cursor"] === undefined) return false;
  return true;
}

export function TwitchClipsResponseFromJSON(json: any): TwitchClipsResponse {
  return TwitchClipsResponseFromJSONTyped(json, false);
}

export function TwitchClipsResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): TwitchClipsResponse {
  if (json == null) {
    return json;
  }
  return {
    state: TwitchClipsStateFromJSON(json["state"]),
    twitchLogin: json["twitch_login"],
    clips: (json["clips"] as Array<any>).map(TwitchClipSummaryFromJSON),
    cursor: json["cursor"],
  };
}

export function TwitchClipsResponseToJSON(json: any): TwitchClipsResponse {
  return TwitchClipsResponseToJSONTyped(json, false);
}

export function TwitchClipsResponseToJSONTyped(
  value?: TwitchClipsResponse | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    state: TwitchClipsStateToJSON(value["state"]),
    twitch_login: value["twitchLogin"],
    clips: (value["clips"] as Array<any>).map(TwitchClipSummaryToJSON),
    cursor: value["cursor"],
  };
}